moving NEAT generators and utilities into language-plutus-core so that they can be used from metatheory
